requirements.txt
python-dotenv~=0.19
python-telegram-bot~=13.7
Pillow~=8.4.0
pymongo~=3.12.1